# src/main/resources/swagger-api.yaml
swagger: "2.0"
info:
  description: "This is a sample server.  This API definition was built using Swagger Editor at [https://editor.swagger.io/](https://editor.swagger.io/). \nFor this sample, you can use the api key `special-key` to test the authorization filters."
  version: "1.0.0"
  title: "Outage Demo"
#  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "outagedevteam@ameren.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "aws.ameren.com"
basePath: "/outage/v1"
...
schemes:
- "https"
paths:
...
  /stat/{account}:
    get:
      tags:
      - "outage"
      summary: "Process a new request for outage status from a customer"
      description: ""
      operationId: "statusRequest"
#      consumes:
#      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "account"
        required: true
        type: string
#        schema: 
#          $ref: "#/definitions/BillAccount"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/OutageStatus"
#      security:
#      - alerts_auth:
#        - "read:alerts"
...
securityDefinitions:
  alerts_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:alerts: "post responses to alerts"
      read:alerts: "read alerts"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  OutageStatus:
    title: "Status of a customer outage"
    description: ""
    properties:
      isPowerOn:
        type: boolean
      account:
        $ref: "#/definitions/BillAccount"
      state: 
        type: string
        enum: ["IL","MO"]
      postalCode:
        type: string
      houseNumber:
        type: string
      outageDateTime:
        type: string
        format: date-time
      esrt:
        title: Estimated Safe Restoration Time
        type: string
        format: date-time
      customersImpacted:
        title: Number of customers affected by outage
        type: integer
        format: int32
      restoredDateTime:
        type: string
        format: date-time
      reportedNotProcessed:
        title: outage report received, but not sent to OMS
        type: boolean
      lastUpdatedDateTime:
        type: string
        format: date-time
      reportAlert:
        title: ??
        type: boolean
      reported:
        title: Has this customer alread reported
        type: boolean
  BillAccount:
    title: "Bill Account Number of customer responding"
    type: "string"
    minLength: 10
    maxLength: 10
    example: "0100500104"
...